Add UnreadWidgetConfigurationFragment
