+++ app/k9mail/src/main/java/com/fsck/k9/widget/unread/UnreadWidgetConfigurationFragment.kt
@@ -1,0 +1,183 @@
+package com.fsck.k9.widget.unread
+
+import android.app.Activity
+import android.appwidget.AppWidgetManager
+import android.content.Intent
+import android.os.Bundle
+import android.view.Menu
+import android.view.MenuInflater
+import android.view.MenuItem
+import android.widget.Toast
+import androidx.core.os.bundleOf
+import androidx.preference.CheckBoxPreference
+import androidx.preference.Preference
+import com.fsck.k9.Preferences
+import com.fsck.k9.R
+import com.fsck.k9.activity.ChooseAccount
+import com.fsck.k9.search.SearchAccount
+import com.fsck.k9.ui.choosefolder.ChooseFolderActivity
+import com.takisoft.preferencex.PreferenceFragmentCompat
+import org.koin.android.ext.android.inject
+
+class UnreadWidgetConfigurationFragment : PreferenceFragmentCompat() {
+    private val preferences: Preferences by inject()
+    private val repository: UnreadWidgetRepository by inject()
+    private val unreadWidgetUpdater: UnreadWidgetUpdater by inject()
+
+    private var appWidgetId: Int = AppWidgetManager.INVALID_APPWIDGET_ID
+    private lateinit var unreadAccount: Preference
+    private lateinit var unreadFolderEnabled: CheckBoxPreference
+    private lateinit var unreadFolder: Preference
+
+    private var selectedAccountUuid: String? = null
+    private var selectedFolder: String? = null
+
+    override fun onCreatePreferencesFix(savedInstanceState: Bundle?, rootKey: String?) {
+        setHasOptionsMenu(true)
+        setPreferencesFromResource(R.xml.unread_widget_configuration, rootKey)
+
+        appWidgetId = arguments?.getInt(ARGUMENT_APP_WIDGET_ID) ?: error("Missing argument '$ARGUMENT_APP_WIDGET_ID'")
+
+        unreadAccount = findPreference(PREFERENCE_UNREAD_ACCOUNT)!!
+        unreadAccount.onPreferenceClickListener = Preference.OnPreferenceClickListener {
+            val intent = Intent(requireContext(), ChooseAccount::class.java)
+            startActivityForResult(intent, REQUEST_CHOOSE_ACCOUNT)
+            false
+        }
+
+        unreadFolderEnabled = findPreference(PREFERENCE_UNREAD_FOLDER_ENABLED)!!
+        unreadFolderEnabled.onPreferenceChangeListener = Preference.OnPreferenceChangeListener { _, _ ->
+            unreadFolder.summary = getString(R.string.unread_widget_folder_summary)
+            selectedFolder = null
+            true
+        }
+
+        unreadFolder = findPreference(PREFERENCE_UNREAD_FOLDER)!!
+        unreadFolder.onPreferenceClickListener = Preference.OnPreferenceClickListener {
+            val intent = ChooseFolderActivity.buildLaunchIntent(
+                context = requireContext(),
+                accountUuid = selectedAccountUuid!!,
+                showDisplayableOnly = true
+            )
+            startActivityForResult(intent, REQUEST_CHOOSE_FOLDER)
+            false
+        }
+    }
+
+    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
+        if (resultCode == Activity.RESULT_OK && data != null) {
+            when (requestCode) {
+                REQUEST_CHOOSE_ACCOUNT -> {
+                    val accountUuid = data.getStringExtra(ChooseAccount.EXTRA_ACCOUNT_UUID)!!
+                    handleChooseAccount(accountUuid)
+                }
+                REQUEST_CHOOSE_FOLDER -> {
+                    val folderServerId = data.getStringExtra(ChooseFolderActivity.RESULT_SELECTED_FOLDER)!!
+                    val folderDisplayName = data.getStringExtra(ChooseFolderActivity.RESULT_FOLDER_DISPLAY_NAME)!!
+                    handleChooseFolder(folderServerId, folderDisplayName)
+                }
+            }
+        }
+        super.onActivityResult(requestCode, resultCode, data)
+    }
+
+    private fun handleChooseAccount(accountUuid: String) {
+        val userSelectedSameAccount = accountUuid == selectedAccountUuid
+        if (userSelectedSameAccount) {
+            return
+        }
+
+        selectedAccountUuid = accountUuid
+        selectedFolder = null
+        unreadFolder.summary = getString(R.string.unread_widget_folder_summary)
+        if (SearchAccount.UNIFIED_INBOX == selectedAccountUuid) {
+            handleSearchAccount()
+        } else {
+            handleRegularAccount()
+        }
+    }
+
+    private fun handleSearchAccount() {
+        if (SearchAccount.UNIFIED_INBOX == selectedAccountUuid) {
+            unreadAccount.setSummary(R.string.unread_widget_unified_inbox_account_summary)
+        }
+        unreadFolderEnabled.isEnabled = false
+        unreadFolderEnabled.isChecked = false
+        unreadFolder.isEnabled = false
+        selectedFolder = null
+    }
+
+    private fun handleRegularAccount() {
+        val selectedAccount = preferences.getAccount(selectedAccountUuid)
+        val accountDescription: String? = selectedAccount.description
+        val summary = if (accountDescription.isNullOrEmpty()) selectedAccount.email else accountDescription
+
+        unreadAccount.summary = summary
+        unreadFolderEnabled.isEnabled = true
+        unreadFolder.isEnabled = true
+    }
+
+    private fun handleChooseFolder(folderServerId: String, folderDisplayName: String) {
+        selectedFolder = folderServerId
+        unreadFolder.summary = folderDisplayName
+    }
+
+    override fun onCreateOptionsMenu(menu: Menu, inflater: MenuInflater) {
+        inflater.inflate(R.menu.unread_widget_option, menu)
+    }
+
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        return when (item.itemId) {
+            R.id.done -> {
+                if (validateWidget()) {
+                    updateWidgetAndExit()
+                }
+                true
+            }
+            else -> super.onOptionsItemSelected(item)
+        }
+    }
+
+    private fun validateWidget(): Boolean {
+        if (selectedAccountUuid == null) {
+            Toast.makeText(requireContext(), R.string.unread_widget_account_not_selected, Toast.LENGTH_LONG).show()
+            return false
+        } else if (unreadFolderEnabled.isChecked && selectedFolder == null) {
+            Toast.makeText(requireContext(), R.string.unread_widget_folder_not_selected, Toast.LENGTH_LONG).show()
+            return false
+        }
+        return true
+    }
+
+    private fun updateWidgetAndExit() {
+        val configuration = UnreadWidgetConfiguration(appWidgetId, selectedAccountUuid!!, selectedFolder)
+        repository.saveWidgetConfiguration(configuration)
+
+        unreadWidgetUpdater.update(appWidgetId)
+
+        // Let the caller know that the configuration was successful
+        val resultValue = Intent()
+        resultValue.putExtra(AppWidgetManager.EXTRA_APPWIDGET_ID, appWidgetId)
+
+        val activity = requireActivity()
+        activity.setResult(Activity.RESULT_OK, resultValue)
+        activity.finish()
+    }
+
+    companion object {
+        private const val ARGUMENT_APP_WIDGET_ID = "app_widget_id"
+
+        private const val PREFERENCE_UNREAD_ACCOUNT = "unread_account"
+        private const val PREFERENCE_UNREAD_FOLDER_ENABLED = "unread_folder_enabled"
+        private const val PREFERENCE_UNREAD_FOLDER = "unread_folder"
+
+        private const val REQUEST_CHOOSE_ACCOUNT = 1
+        private const val REQUEST_CHOOSE_FOLDER = 2
+
+        fun create(appWidgetId: Int): UnreadWidgetConfigurationFragment {
+            return UnreadWidgetConfigurationFragment().apply {
+                arguments = bundleOf(ARGUMENT_APP_WIDGET_ID to appWidgetId)
+            }
+        }
+    }
+}
